Move Point to point submodule
